feat: uuid 및 venderSeq 바인딩

+++ src/main/java/susussg/pengreenlive/user/Service/UserService.java
@@ -5,7 +5,7 @@ import susussg.pengreenlive.user.dto.UpdateUserFormDTO;
 
 public interface UserService {
     String getUserNmByUUID(String uuid);
-    String getChannelNmByVendorSeq(int vendorSeq);
+    String getChannelNmByVendorSeq(Long vendorSeq);
     String getUserAddressByUUID(String uuid);
 
     void updateUserInfo(UpdateUserFormDTO user);
